backend/Chatbot/requirements.txt
python-telegram-bot
python-socketio
requests
aiohttp
python-dotenv
